guava/backend/api_gateway/requirements.txt
# API Gateway Requirements
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0


